@@ -3,9 +3,10 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "build:watch": "ng build --watch",
+    "build:watch": "ng build --watch --prod=false",
     "ng": "ng",
     "start": "ng serve",
+    "start:debug": "ng serve -c=debug",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -13,45 +14,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
-    "core-js": "^2.5.5",
+    "@angular/common": "^6.0.4",
+    "@angular/compiler": "^6.0.4",
+    "@angular/core": "^6.0.4",
+    "@angular/forms": "^6.0.4",
+    "@angular/http": "^6.0.4",
+    "@angular/platform-browser": "^6.0.4",
+    "@angular/platform-browser-dynamic": "^6.0.4",
+    "@angular/router": "^6.0.4",
+    "core-js": "^2.5.7",
     "enhanced-resolve": "^3.4.1",
     "fast-json-patch": "^2.0.6",
     "json-server": "^0.11.2",
     "lodash": "^4.17.10",
+    "ngx-cookie-service": "^1.0.10",
+    "object-hash": "^1.3.0",
     "q": "^1.5.1",
-    "rxjs": "^5.5.10",
+    "rxjs": "^6.2.1",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@types/jasmine": "^2.8.6",
-    "@types/lodash": "^4.14.108",
-    "@types/node": "^6.0.106",
-    "codelyzer": "~4.0.0",
+    "@angular-devkit/build-angular": "^0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.4",
+    "@types/jasmine": "^2.8.8",
+    "@types/lodash": "^4.14.109",
+    "@types/node": "^6.0.112",
+    "@types/object-hash": "^1.2.0",
+    "codelyzer": "^4.3.0",
     "concurrently": "^3.5.1",
+    "jasmine": "^3.1.0",
     "jasmine-core": "^2.99.1",
     "jasmine-spec-reporter": "~3.2.0",
     "karma": "~1.4.1",
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-hue-reporter": "0.0.4",
-    "karma-jasmine": "^1.1.1",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "node-sass": "^4.9.0",
     "protractor": "~5.1.0",
     "ts-mockito": "^2.3.0",
     "ts-node": "~2.0.0",
     "tslint": "~5.7.0",
-    "typescript": "2.6.2"
+    "typescript": "~2.7.2"
   }
 }
