@@ -1,38 +1,82 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-
-import { ApiService } from '../../app-common/api.service';
 import { ActivatedRoute } from '@angular/router';
 import { CRUDRouter } from '../../../api-routes/CRUDRouter';
-import { Project } from '../../../api-models/project';
-import { Subscription } from 'rxjs/Subscription';
-import 'rxjs/add/operator/mergeMap';
+import { Project, Sequence } from '../../../api-models';
+import { ComponentDynamicStates } from '../../app-common/dynamic-state/dynamic-state.component';
+import { DataService } from '../../app-common/data.service';
+import { Observable, Subscription } from 'rxjs/index';
+import { map } from 'rxjs/operators';
+import { BaseDetail } from '../../app-common/base-detail';
+import { CRUD } from '../../app-common/api.service';
+
+
 
 @Component({
   selector: 'app-project-detail',
   templateUrl: './project-detail.component.html',
   styleUrls: ['./project-detail.component.scss']
 })
-export class ProjectDetailComponent implements OnInit, OnDestroy {
-  project: Project;
-  projectSubscription: Subscription;
+export class ProjectDetailComponent extends BaseDetail<Project> {
+  projectState: Observable<ComponentDynamicStates>;
+  private projectIdSubscription: Subscription;
 
   constructor(
-    private api: ApiService,
-    private route: ActivatedRoute
-  ) { }
+    data: DataService,
+    route: ActivatedRoute
+  ) {
+    super(data);
+    this.projectState = this.$state.asObservable();
 
-  ngOnInit() {
-    this.projectSubscription = this.route.params.mergeMap(({projectId}) => {
-       return this.api.getData<Project>(CRUDRouter.getProject, {projectId});
-    }).subscribe(p => this.project = p);
+    this.callConfigurator = (project, cause) => {
+      switch (cause) {
+        case CRUD.CREATE: {
+          return {
+            route: CRUDRouter.newProject
+          }
+        }
+        case CRUD.READ: {
+          return {
+            route: CRUDRouter.getProject,
+            params: { projectId: this.dataItemId }
+          }
+        }
+        case CRUD.UPDATE: {
+          return {
+            route: CRUDRouter.saveProject,
+            params: { projectId: this.dataItemId }
+          }
+        }
+        case CRUD.DELETE: {
+          return {
+            route: CRUDRouter.deleteProject,
+            params: { projectId: this.dataItemId }
+          }
+        }
+      }
+    }
+
+    this.dataItemIdObservableSource = () => {
+      return route.params.pipe(map(({projectId}) => projectId));
+    };
   }
 
-  ngOnDestroy() {
-    this.projectSubscription.unsubscribe();
+  save(form) {
+    if (form.valid) {
+      this.saveDataItem()
+    }
   }
 
-  onPlaceholded(e, d) {
-    console.log(e, d);
+  editProject() {
+    this.$state.next(ComponentDynamicStates.EDITING);
   }
 
+  onTopSequenceChange(topSequence: Sequence) {
+    if (!this.dataItem.topSequenceId || this.dataItem.topSequenceId !== topSequence.id) {
+      this.dataItem = {
+        ...this.dataItem,
+        topSequence
+      };
+      this.saveDataItem();
+    }
+  }
 }
