@@ -5,11 +5,17 @@ import { FormsModule } from '@angular/forms';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ApiService } from '../../app-common/api.service';
 import { Project } from '../../../api-models';
-import { Observable } from 'rxjs/Observable';
+import { Observable, of } from 'rxjs';
+import { AppCommonModule } from '../../app-common/app-common.module';
+import { InputsModule } from '../../inputs/inputs.module';
+import { ProjectThumbnailComponent } from '../project-thumbnail/project-thumbnail.component';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { DataService } from '../../app-common/data.service';
+
 
 class MockApi {
   getData() {
-    return Observable.of(<Project[]>[{
+    return of(<Project[]>[{
       name: 'test project',
       description: 'test project description',
     }])
@@ -26,14 +32,18 @@ describe('ProjectListComponent', () => {
     TestBed.configureTestingModule({
       imports: [
         FormsModule,
-        RouterTestingModule
+        RouterTestingModule,
+        InputsModule
       ],
       declarations: [
-        ProjectListComponent
+        ProjectListComponent,
+      ],
+      schemas: [
+        NO_ERRORS_SCHEMA
       ],
       providers: [
         {
-          provide: ApiService,
+          provide: DataService,
           useClass: MockApi
         }
       ]
