@@ -1,5 +1,5 @@
-import { OrderableBase } from './base';
 import { AssetContentTypes, AssetTypes } from './enums/asset.enums';
+import { OrderableBase } from './orderableBase';
 export declare class AssetBase extends OrderableBase {
     assetType?: AssetTypes;
     contentType?: AssetContentTypes;
