@@ -1,12 +1,6 @@
-export declare enum ItemLifeCycleTypes {
-    ONE_ONE = "1:1",
-    N_ONE = "N:1",
-    ONE_N = "1:N",
-    N_N = "N:N",
-}
-export declare enum QuantityOrder {
-    NONE = "none",
-    ONE = "one",
-    MULTIPLE = "multiple",
-    TWO_DIMENSIONAL = "2d",
+export enum QuantityOrder {
+    NONE = 'none',
+    ONE = 'one',
+    MULTIPLE = 'multiple',
+    TWO_DIMENSIONAL = '2d'
 }
