@@ -1,6 +1,6 @@
-export declare enum StepTypes {
-    ITEM_REF = "item_reference",
-    TASK_REF = "task_reference",
-    LOGICAL_ITEM = "logical_item",
-    SEQUENCE_REF = "sequence_reference",
+export enum StepTypes {
+    QUESTION_REF = 'question_reference',
+    TASK_REF = 'task_reference',
+    LOGICAL_ITEM = 'logical_item',
+    SEQUENCE_REF = 'sequence_reference'
 }
