@@ -4,6 +4,6 @@ import { SequenceHeader } from './sequenceHeader';
 export declare class Sequence extends Base {
     header?: SequenceHeader;
     steps?: Step[];
+    stepIds?: number[];
     referencedBySteps?: Sequence[];
-    sortSteps?(): void;
 }
